+++ src/main/java/org/cookbookgeeks/webkochbuch/dao/IngredientDao.java
@@ -26,7 +26,7 @@ import org.cookbookgeeks.webkochbuch.domain.Ingredient;
  * @author Nils Sommer
  *
  */
-public class IngredientDao extends GenericDao<Ingredient, Integer> {
+public class IngredientDao extends GenericDao<Ingredient, Long> {
 	
 	/**
 	 * Standard constructor.
